fix: Split PhaseLogCard specs for lint budget
The title-line tests pushed PhaseLogCard.spec.tsx past 500 counted
lines. That added one max-lines finding and failed the ESLint budget.
Split those cases into a second spec and share fixtures so CI stays
inside the existing ratchet.
